requirements.txt
Django==2.2.1
djangorestframework==3.9.4
omdb==0.10.1